Применяет createSlice
