+++ src/store/index.ts
@@ -1,7 +1,7 @@
 import { configureStore } from '@reduxjs/toolkit';
-import { reducer } from './reducer';
+import { offersSlice } from './slice/offers';
 
 
 export const store = configureStore({
-  reducer,
+  reducer: offersSlice.reducer,
 });
